test/ rearrange tests
