+++ tests/unit/test_scraping/test_review_fetcher.py
@@ -1,9 +1,7 @@
 import pytest
 import requests
-from unittest.mock import Mock, patch, MagicMock
-from typing import List, Optional, Dict, Any
+from unittest.mock import Mock, patch
 
-from src.scraping.fetcher import Fetcher
 from src.scraping.review_fetcher import ReviewFetcher
 
 
@@ -164,52 +162,4 @@ class TestReviewFetcher:
         assert hasattr(self.fetcher, 'fetch_multiple_pages')
         assert callable(self.fetcher.fetch_multiple_pages)
 
-
-class TestReviewFetcherIntegration:
-    """Integration tests for ReviewFetcher (these require network access)"""
-    
-    def setup_method(self):
-        """Set up test fixtures"""
-        self.fetcher = ReviewFetcher("https://httpbin.org")  # Using httpbin for reliable testing
-    
-    @pytest.mark.integration
-    def test_fetch_page_real_request(self):
-        """Test fetch_page with a real HTTP request"""
-        # Skip if running without internet access
-        pytest.importorskip("requests")
-        
-        try:
-            result = self.fetcher.fetch_page("https://httpbin.org/html")
-            assert "<html>" in result
-            assert len(result) > 0
-        except requests.RequestException:
-            pytest.skip("Network not available for integration test")
-    
-    @pytest.mark.integration
-    def test_fetch_multiple_pages_real_requests(self):
-        """Test fetch_multiple_pages with real HTTP requests"""
-        pytest.importorskip("requests")
-        
-        urls = [
-            "https://httpbin.org/html",
-            "https://httpbin.org/json"
-        ]
-        
-        try:
-            result = self.fetcher.fetch_multiple_pages(urls)
-            assert len(result) == 2
-            assert all(len(content) > 0 for content in result.values())
-        except requests.RequestException:
-            pytest.skip("Network not available for integration test")
-
-
-class TestReviewFetcherAbstractMethods:
-    """Test that ReviewFetcher implements required abstract methods"""
-    
-    def setup_method(self):
-        """Set up test fixtures"""
-        self.fetcher = ReviewFetcher("https://example.com")
-    
-
-
